requirements.txt
Django==4.2.4
django-jet-reboot==1.3.5
psycopg2==2.9.7
django-allauth==0.59.0
Pillow==10.0.0
django-cors-headers==4.2.0
django-nested-admin==4.0.2
django-treebeard==4.7
django-mptt==0.15.0
python-dotenv==1.0.0